feat(tool-stack): move cli-plugin-stack into tool-stack
