--- packages/@code-like-a-carpenter/cli-plugin-stack-name/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@code-like-a-carpenter/cli-plugin-stack-name",
-  "description": "Converts an input string to a stack name",
-  "dependencies": {
-    "@code-like-a-carpenter/cli-core": "*",
-    "@code-like-a-carpenter/tooling-common": "*"
-  },
-  "author": "Ian Remmel <1182361+ianwremmel@users.noreply.github.com> (https://www.ianwremmel.com)",
-  "bugs": "https://www.github.com/code-like-a-carpenter/workbench/issues",
-  "engines": {
-    "node": "20.x"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/types/index.d.ts",
-      "development": "./src/index.ts",
-      "import": "./dist/esm/index.mjs"
-    },
-    "./package.json": "./package.json"
-  },
-  "homepage": "https://www.github.com/code-like-a-carpenter/workbench/tree/main/packages/@code-like-a-carpenter/cli-plugin-stack-name",
-  "license": "MIT",
-  "repository": "git@github.com:code-like-a-carpenter/workbench.git",
-  "types": "dist/types",
-  "publishConfig": {
-    "access": "public"
-  },
-  "bin": "./cli.mjs",
-  "releaseAll": 1
-}
