--- packages/@code-like-a-carpenter/cli-plugin-stack-name/src/index.ts
@@ -1,16 +1,0 @@
-import {definePlugin} from '@code-like-a-carpenter/cli-core';
-import {getStackName} from '@code-like-a-carpenter/tooling-common';
-
-export default definePlugin((yargs) => {
-  yargs.command({
-    builder: (y) =>
-      y.positional('projectName', {
-        demandOption: true,
-        type: 'string',
-      }),
-    command: 'stack-name <projectName>',
-    async handler(argv) {
-      console.log(getStackName(argv.projectName));
-    },
-  });
-});
