--- packages/@code-like-a-carpenter/cli-plugin-stack-name/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist/types",
-    "rootDir": "./src"
-  },
-  "extends": "../../../tsconfig.references.json",
-  "include": ["src"],
-  "references": [
-    {
-      "path": "../cli-core"
-    },
-    {
-      "path": "../tooling-common"
-    }
-  ]
-}
