+++ packages/@code-like-a-carpenter/cli/package.json
@@ -19,7 +19,6 @@
     "plugins": [
       "@code-like-a-carpenter/cli-plugin-example",
       "@code-like-a-carpenter/cli-plugin-proxy",
-      "@code-like-a-carpenter/cli-plugin-stack-name",
       "@code-like-a-carpenter/tool-stack",
       "@code-like-a-carpenter/tool-tool",
       "@code-like-a-carpenter/tooling-deps"
@@ -29,7 +28,6 @@
     "@code-like-a-carpenter/cli-core": "*",
     "@code-like-a-carpenter/cli-plugin-example": "*",
     "@code-like-a-carpenter/cli-plugin-proxy": "*",
-    "@code-like-a-carpenter/cli-plugin-stack-name": "*",
     "@code-like-a-carpenter/tool-stack": "*",
     "@code-like-a-carpenter/tool-tool": "*",
     "@code-like-a-carpenter/tooling-deps": "*"
