+++ packages/@code-like-a-carpenter/tool-stack/tsconfig.json
@@ -7,7 +7,13 @@
   "include": ["src"],
   "references": [
     {
+      "path": "../assert"
+    },
+    {
       "path": "../cli-core"
+    },
+    {
+      "path": "../tooling-common"
     }
   ]
 }
